refactor(ui): split App.tsx into 5 modules
